@@ -3,7 +3,7 @@ package backtraceio.library;
 import static org.junit.Assert.assertEquals;
 
 import android.net.Uri;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
