@@ -3,7 +3,7 @@ package backtraceio.library.common;
 import android.Manifest;
 import android.content.Context;
 import android.content.pm.PackageManager;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 
 /***
  * Helper class for checking permissions
