@@ -2,7 +2,7 @@ package backtraceio.backtraceio;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.system.ErrnoException;
 import android.system.Os;
 import android.util.Log;
